scripts/requirements.txt
# librosa core
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0

# MIDI generation and SysEx capture
mido>=1.3.0
python-rtmidi>=1.5.0
python-rtmidi>=1.4.9

# OSC bridge client (iron-static-bridge.amxd)
python-osc>=1.8.3

# Audio-to-MIDI transcription (audio_to_midi.py)
basic-pitch>=0.3.0      # Spotify pitch tracker — ONNX backend, no GPU needed
demucs>=4.0.0           # Meta stem separator (drums/bass/other/vocals)

# Ableton bridge (ableton_push.py)
python-hcl2>=4.3.0      # HCL config file parser for session templates

# LLM integration (automated workflows — not for interactive Copilot sessions)
google-genai>=1.0.0

# GCS large-file sync (gcs_sync.py)
google-cloud-storage>=2.16.0

# Social media content generation (generate_promo_image.py, generate_caption.py)
# Note: render_waveform_video.py uses ffmpeg (system binary, brew install ffmpeg)

# General utilities
argparse  # stdlib in Python 3.2+
